[core] test
